Add fireRule() method to Rule interface.
