+++ src/main/java/com/deloitte/elrr/aggregator/rules/Rule.java
@@ -5,5 +5,7 @@ import com.yetanalytics.xapi.model.Statement;
 
 public interface Rule {
 
+  boolean fireRule(Statement statement);
+
   void processRule(Person person, Statement statement) throws Exception;
 }
